Refactoring recording unit form bean
