+++ src/main/java/fr/siamois/bean/Converter/PersonConverter.java
@@ -1,19 +1,15 @@
-package fr.siamois.bean.Person;
+package fr.siamois.bean.Converter;
 
 import fr.siamois.models.auth.Person;
 import fr.siamois.services.auth.PersonService;
 import jakarta.faces.component.UIComponent;
 import jakarta.faces.context.FacesContext;
 import jakarta.faces.convert.Converter;
-import jakarta.faces.convert.FacesConverter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 
-import javax.faces.bean.ManagedBean;
 
-@ManagedBean
 @Component
 @Slf4j
 public class PersonConverter implements Converter<Person> {
